lots of refactoring for new activity manager

+++ src/wbs/clients/apn/chat/user/core/model/ChatUserDaoMethods.java
@@ -17,7 +17,7 @@ interface ChatUserDaoMethods {
 			ChatRec chat,
 			NumberRec number);
 
-	int countOnline (
+	Long countOnline (
 			ChatRec chat,
 			ChatUserType type);
